+++ code/test/Templates.Test/BuildTemplatesTests/Uwp/BuildMVVMLightProjectTests.cs
@@ -11,7 +11,7 @@ using Microsoft.Templates.Core.Extensions;
 
 using Xunit;
 
-namespace Microsoft.Templates.Test
+namespace Microsoft.Templates.Test.Build.Uwp
 {
     [Collection("BuildTemplateTestCollection")]
     public class BuildMVVMLightProjectTests : BaseGenAndBuildTests
