+++ code/test/Templates.Test/VBCodeAnalysis/VBStyleCollection.cs
