+++ code/test/Templates.Test/VBCodeAnalysis/VBStyleGenerationTestsFixture.cs
