+++ code/test/Templates.Test/VBCodeAnalysis/VBStyleProjectGenerationTests.cs
