+++ code/test/Templates.Test/bat/RestoreAndBuildXamarin.bat
