+++ code/test/Templates.Test/bat/RunWackTest.bat
