+++ code/test/Templates.Test/bat/Uwp/RestoreAndBuild.bat
