+++ code/test/Templates.Test/bat/Uwp/RestoreAndBuildAppx.bat
