+++ code/test/Templates.Test/bat/Uwp/RunTests.bat
