+++ code/test/Templates.Test/bat/Wpf/RestoreAndBuild.bat
