+++ code/test/Templates.Test/bat/Wpf/RestoreAndBuildWithMsix.bat
